Step 11 - Add test case of multiple delimiters
This is a really curious case. With the current code, we don't need to add more production code to fulfill the requirements. 
This happens because our last refactor. This refactor introduces more behavior than we need, but on the other side, we got a easier to understand and simpler code.  
In this case, I just cover this behavior with tests to fix and document it but nothing else